Add Remark Function
Added the ability to add optional remarks to people
in the address book and edit them when required.

This was done by adding the following classes:
* RemarkCommand - main code for adding remarks to person
* RemarkCommandParser - makes sense of user args
* Remark - acts as field for Person

The person class was also altered to accept remark as a field.
Consequently, all code utilizing the person constructor had to be
changed as well.

Lastly, automated tests were added to ensure that the new function
behaves as expected.

The function was added to add functionality to the address book for its
users.

The secondary objective of the implementation of this function was to
understand the code architecture and the correct process behind
developing the codebase and adding commands.