Bugfix: Find (Reflection and Studio)
Currently we are able to use the find command to group students based
on reflection and studio classes.

However, the current method searches for substrings, meaning anyone
searching for studio class S1 would end up seeing any studio which has
"S1" in it, e.g. S11, S12 etc. This would make searching for the
students in specific classes unnecessarily troublesome.

Let's change the way find works and search for the exact class when
presented with the prefix s/ or r/.

By searching for the exact class, we can be sure that the user will
be able to see the intended class only.