feat: update export command to export exam name and scores to csv
users should be able to export selected persons exam scores along with
the other details stored when they use the export command.

add,
* read exams array from unfiltered json file (whole addressbook) to
obtain the exam names as headings for the csv file. add "Exam:" to the
front of the names so that headings will be displayed as
Exam:[NAME_OF_EXAM] in csv file for ease of future exports. this also
prevents errors if any of the exam names match the other fieldnames.
* edit jsonTree for personsArray to change the nodes such that
the examScores node and its children are removed and new key-value pairs
[exam_name]: [score] is added for ease of writing jsonTree to csv
schema built