Improve SLAP for find functionality
Currently, the find command's functionality has no issue. However as
@Zer0Legion pointed out, the functions can be further abstracted.

We need to make the changes as he has suggested to improve the
readability of the code in general.

Let's incorporate SLAP into some nested functions to reduce the amount
of nested code.

In doing so, we can promote code clarity, maintainability, and
reusability, where functions and modules are focused on completing one
task at a time.