Rename boolean methods correctly
Currently the find functionality works as intended.
However, two Boolean methods in the
`PersonDetailContainsKeywordAndExamPredicate.java` file were incorrectly
named as they do not sound like a Boolean.

This has to be changed immediately as it violates the coding standards
for style.

Let's rename these methods correctly, following the coding conventions.

By doing so, we can ensure that our code keeps to the expected
coding standard.