Update chars accepted by Name
Currently, the name accepts alphanumerical characters, spaces, and
forward slashes. This will allow the majority of names to be added
into the app.

However, it has come to light that there are names which would contain
other characters that would throw errors if users attempt to add these
names in, which is problematic since some special characters like `.`
and `,` are common.

Let's allow several other characters in the regex for names.

In doing so, we can ensure that most common names will be successfully
added into the application.