Refactor Find Functionality
Currently, the find command holds a predicate, and creates another
predicate based on whether exam scores are needed.

However, this may be messy due to the fact that the predicate
objects are created in different places despite serving the same
purpose.

Let's refactor the creation of the predicate objects to ensure
that the objects follow a more logical and organized flow.

In doing so, the code is both more readable and scalable.