Fix NullPointerException when clicking UI
Currently, the application throws a NullPointerException when items in the Exam List and the Person List are clicked on in the User Interface.
This is due to bad implementation practices, as I had put a band aid on this problem by setting the selectionModel of both lists to null to prevent the selection of exams in the list..
While this implementation does not affect the use of the application, Lets update the implementation to use a proper "NoSelectionModel".
This will prevent further NullPointerExceptions from being thrown in the application.