Clarify email and copy related docs and messages
Currently, we were able to observe from the PE dry run that there is
a possibility for testers to perceive from the UG and the error
messages that there may be bugs present despite the code working as
intended.

This is an issue because it indicates that the documentation is not
clear enough to convey the exact reason behind the current behavior of
the code (e.g. semicolons after each copied email after using the
`copy` command, "seemingly valid" emails being rejected).

Let's ramp up the clarity in both the UG and constraint messages
pertaining emails.

In doing so, future users will have a much clearer idea of how the code
is supposed to behave, and will not mistake correct behaviors as bugs.