UG improvements
Currently, as the testers from the PE-D have pointed out, the UG has
room for improvement in regards to the clarity of some of the commands
as well as the omission of usage information for some prefix.

This has to be fixed because future users would likely be unclear in
the purpose behind some of the features of the application.

Let's revamp the UG to include the missing use cases and improve the
clarity of the user guide.

In doing so, future users will have a better idea of the expected
behavior of the app and the rationale of why the behavior of the app
was designed as such.