Update Manual Testing Section in DG
Currently, the manual testing section does not include instructions for
functionalities currently implemented.

We need to keep this updated to allow future developers to check
each functionality manually to ensure the code behaves as intended.

Let's update the Manual Testing instructions to include comprehensive
instructions for the following commands:
* `help`
* `clear`
* `find`
* `copy`
* `exit`

In doing so, these commands can be tested with the assurance that
all aspects of the behaviours of these functions are covered.