Bug Fix: Find Command Parser
Currently, the find command does not check that there are duplicate
prefixes present in the input provided by the user.

This is problematic because it allows the find command to accept
multiple duplicate prefixes despite the command being documented to
only accept one single prefix.

An example is `find n|alice n|bob`
* Expected behaviour: throw an error indicating more than one prefix
is used.
* Actual behaviour: finds persons with names containing bob.

Because this behaviour differs from the user guide, does not handle the
error of duplicate prefixes, and behaves different from normal
expectations, the following error can be classified as a functionality
bug, and not a feature flaw.

Let's rectify this problem by using the in-built duplicate verifier to
properly handle the issue of duplicate prefixes.

Doing so not only solves the bug, but also does so in a way that does
not implement any new functionality into the code.