Reorganize DG manual testing
Currently, the DG manual testing section is not organized according to
the order of the UG.

However, we want to maintain a level of standardization that gives
the impression that all documentation was written by one person.

Let's reorder the manual testing section to maintain a standard order.

By keeping the structure standardized, the DG will not only appear more
organized, but it will help the future developers in navigating through
the DG.