Edit DG and diagrams for find implementation
Currently, the find implementation documented in the DG is outdated,
as it details the previous implementation before the restructuring.
Furthermore, the diagrams are no longer accurate with the addition of
the exam score compatible functionality not being represented.

We need to update this promptly since we want the developer guide
to accurately represent the most recent implementation to aid in the
future development of the application.

Let's update the find implementation details in the developer guide,
as well as the supporting diagrams.

By keeping the information up to date, developers will be able to
accurately understand the implementation of the find functionality.