Correct minor issues in DG
Currently, minor grammatical errors were spotted in the DG after
the previous commit.

This needs to be corrected as it undermines the professionalism of the
developer guide.

Let's rectify all language issues in the edited portion of the DG.

In maintaining the DG, we can avoid misinterpretation and uphold a high
standard of professionalism.