Edit page breaks in UG
Currently, the page break before the "Adding an Exam" section in the
UG is causing an additional blank page to be generated when the UG
is printed.

This is an issue as it undermines the efforts made to format the UG
properly.

Let's amend the UG to ensure that the printed UG appears as intended.

By doing so, we prevent unwanted issues from printing the user guide.