Update Manual Testing Section of DG
Currently, the manual testing section is formatted differently
since it was written by multiple people.

This needs to be rectified since the manual testing section looks to not
follow any standard, and looks untidy and unstructured.

Let's reformat the manual testing section to follow a similar structure.

In unifying the format, we can ensure that the dev guide looks
professionally made and well organized. Furthermore, the manual testing
section will be easier to understand and follow if everything is
standardized.